+++ src/components/article-page/article-page.tsx
@@ -1,9 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
 import { useParams } from 'react-router-dom';
-import { fetchArticleHtml, fetchArticles } from '../../lib/fetch';
+import { fetchArticles } from '../../lib/fetch';
 import Spinner from '../common/spinner/spinner';
-import sanitizeHtml from 'sanitize-html';
-import style from './article-page.module.css';
+import { Article } from './article/article';
 
 export const ArticlePage = (): JSX.Element => {
   const { articleKey } = useParams();
@@ -22,40 +21,9 @@ export const ArticlePage = (): JSX.Element => {
     return <p>Artikkelen finnes ikke.</p>;
   }
 
-  const articleId = articles
+  const articleId: string = articles
     .filter((article) => article.documentKey === articleKey)
-    .map((article) => article.documentId);
+    .map((article) => article.documentId)[0];
 
-  const {
-    isLoading: isLoadingHtml,
-    isError: isHtmlError,
-    data: articleHtml,
-  } = useQuery({
-    queryKey: ['article', articleId],
-    queryFn: fetchArticleHtml,
-  });
-
-  if (isLoadingHtml) return <Spinner />;
-  if (isHtmlError) return <p>Kunne ikke laste artikkel.</p>;
-
-  const cleanHtml = sanitizeHtml(articleHtml, {
-    allowedTags: sanitizeHtml.defaults.allowedTags.concat(['img']),
-    allowedAttributes: false,
-    transformTags: {
-      img: sanitizeHtml.simpleTransform('img', {
-        referrerpolicy: 'no-referrer',
-      }),
-    },
-  });
-
-  return (
-    <div className="container-sm m-5">
-      <div className="col-12 col-lg-8 mx-auto">
-        <div
-          className={style.article}
-          dangerouslySetInnerHTML={{ __html: cleanHtml }}
-        />
-      </div>
-    </div>
-  );
+  return <Article articleId={articleId} />;
 };
